feat: Add model classes and API routes
Movie class represents abstraction of a single movie. MovieResponse
is response of the API which contains information of movies, current
queried page and total number of pages that will be used in lazy
loading of the movies list. MoviesAPI contains list of API endpoint
routes with parameters that are required to do the queries.